feat(gui): show usage counts for other declarations in selection view
