+++ api-editor/gui/src/features/packageData/selectionView/ClassView.tsx
@@ -8,6 +8,7 @@ import { SectionListViewItem } from './SectionListViewItem';
 import { CompleteButton } from '../../annotations/CompleteButton';
 import { MissingAnnotationButton } from '../../annotations/MissingAnnotationButton';
 import { DataCopyButtons } from '../../annotations/DataCopyButtons';
+import {NonParameterUsageCounts} from "./UsageCounts";
 
 interface ClassViewProps {
     pythonClass: PythonClass;
@@ -45,6 +46,7 @@ export const ClassView: React.FC<ClassViewProps> = function ({ pythonClass }) {
             </Stack>
             <SectionListViewItem title="Superclasses" inputElements={pythonClass.superclasses} />
             <SectionListViewItem title="Decorators" inputElements={pythonClass.decorators} />
+            <NonParameterUsageCounts declaration={pythonClass} />
         </Stack>
     );
 };
