+++ api-editor/gui/src/features/packageData/selectionView/FunctionView.tsx
@@ -15,6 +15,7 @@ import { selectFilter, selectSorter } from '../../ui/uiSlice';
 import { selectUsages } from '../../usages/usageSlice';
 import { MissingAnnotationButton } from '../../annotations/MissingAnnotationButton';
 import { DataCopyButtons } from '../../annotations/DataCopyButtons';
+import {NonParameterUsageCounts} from "./UsageCounts";
 
 interface FunctionViewProps {
     pythonFunction: PythonFunction;
@@ -82,6 +83,8 @@ export const FunctionView: React.FC<FunctionViewProps> = function ({ pythonFunct
                     )}
                 </Stack>
             </Stack>
+
+            <NonParameterUsageCounts declaration={pythonFunction} />
         </Stack>
     );
 };
