+++ api-editor/gui/src/features/packageData/selectionView/ModuleView.tsx
@@ -10,6 +10,7 @@ import { PythonModule } from '../model/PythonModule';
 import { CodeComponent } from 'react-markdown/lib/ast-to-react';
 import { CompleteButton } from '../../annotations/CompleteButton';
 import { DataCopyButtons } from '../../annotations/DataCopyButtons';
+import {NonParameterUsageCounts} from "./UsageCounts";
 
 // See https://github.com/remarkjs/react-markdown#use-custom-components-syntax-highlight
 const CustomCode: CodeComponent = function ({
@@ -101,6 +102,8 @@ export const ModuleView: React.FC<ModuleViewProps> = function ({ pythonModule })
                     </Text>
                 )}
             </Stack>
+
+            <NonParameterUsageCounts declaration={pythonModule} />
         </Stack>
     );
 };
